security_scanner/requirements.txt
functions-framework
google-cloud-containeranalysis